+++ lib/mongoid.rb
@@ -25,12 +25,6 @@ require "mongoid/log_subscriber"
 # the nifty initializers that Mongoid needs.
 if defined?(Rails)
   require "mongoid/railtie"
-else
-  # If we are using any Rack based application then we need the Mongoid rack
-  # middleware to ensure our app is running properly.
-  if defined?(Rack)
-    require "rack/mongoid"
-  end
 end
 
 # add english load path by default