+++ lib/mongoid/relations/many.rb
@@ -21,11 +21,11 @@ module Mongoid
         size == 0
       end
 
-        # Creates a new document on the references many relation. This will
-        # save the document if the parent has been persisted.
-        #
-        # @example Create and save the new document.
-        #   person.posts.create(:text => "Testing")
+      # Creates a new document on the references many relation. This will
+      # save the document if the parent has been persisted.
+      #
+      # @example Create and save the new document.
+      #   person.posts.create(:text => "Testing")
       #
       # @overload create(attributes = nil, options = {}, type = nil)
       #   @param [ Hash ] attributes The attributes to create with.
@@ -40,9 +40,13 @@ module Mongoid
       #
       # @since 2.0.0.beta.1
       def create(attributes = nil, type = nil, &block)
-        doc = build(attributes, type, &block)
-        base.persisted? ? doc.save : raise_unsaved(doc)
-        doc
+        if attributes.is_a?(::Array)
+          attributes.map { |attrs| create(attrs, type, &block) }
+        else
+          doc = build(attributes, type, &block)
+          base.persisted? ? doc.save : raise_unsaved(doc)
+          doc
+        end
       end
 
       # Creates a new document on the references many relation. This will
@@ -66,9 +70,13 @@ module Mongoid
       #
       # @since 2.0.0.beta.1
       def create!(attributes = nil, type = nil, &block)
-        doc = build(attributes, type, &block)
-        base.persisted? ? doc.save! : raise_unsaved(doc)
-        doc
+        if attributes.is_a?(::Array)
+          attributes.map { |attrs| create!(attrs, type, &block) }
+        else
+          doc = build(attributes, type, &block)
+          base.persisted? ? doc.save! : raise_unsaved(doc)
+          doc
+        end
       end
 
       # Find the first document given the conditions, or creates a new document