+++ spec/mongoid/findable_spec.rb
@@ -128,6 +128,20 @@ describe Mongoid::Findable do
     end
   end
 
+  [ :first, :one ].each do |method|
+
+    describe "##{method}" do
+
+      let!(:person) do
+        Person.create
+      end
+
+      it "returns the first matching document" do
+        expect(Person.send(method)).to eq(person)
+      end
+    end
+  end
+
   describe ".first_or_create" do
 
     context "when the document is found" do
@@ -190,7 +204,7 @@ describe Mongoid::Findable do
         end
 
         it "calls the block" do
-          expect(person.pets).to be_true
+          expect(person.pets).to be true
         end
       end
     end
@@ -266,7 +280,7 @@ describe Mongoid::Findable do
         end
 
         it "calls the block" do
-          expect(person.pets).to be_true
+          expect(person.pets).to be true
         end
       end
     end