+++ spec/mongoid/matchable/ne_spec.rb
@@ -11,14 +11,35 @@ describe Mongoid::Matchable::Ne do
     context "when the values are not equal" do
 
       it "returns true" do
-        expect(matcher.matches?("$ne" => "second")).to be_true
+        expect(matcher.matches?("$ne" => "second")).to be true
       end
     end
 
     context "when the values are equal" do
 
       it "returns false" do
-        expect(matcher.matches?("$ne" => "first")).to be_false
+        expect(matcher.matches?("$ne" => "first")).to be false
+      end
+    end
+
+    context "when the value is an array" do
+
+      let(:array_matcher) do
+        described_class.new([ "first" ])
+      end
+
+      context "when the value is in the array" do
+
+        it "returns false" do
+          expect(array_matcher.matches?("$ne" => "first")).to be false
+        end
+      end
+
+      context "when the value is not in the array" do
+
+        it "returns true" do
+          expect(array_matcher.matches?("$ne" => "second")).to be true
+        end
       end
     end
   end